requirements.txt
opencv-python
numpy
torch
torchvision
pillow
ftfy
regex
tqdm
CLIP @ git+https://github.com/openai/CLIP.git
ultralytics
